Disabled getting descriptions (function) from UniProt web ser. Merged cBioPortal accessors. etc. cleanup
