+++ pom.xml
@@ -19,6 +19,15 @@
 		<cpath-client.version>7.0.0-SNAPSHOT</cpath-client.version>
 	</properties>
 
+	<pluginRepositories>
+		<pluginRepository>
+			<id>apache-snapshots</id>
+			<name>Apache snapshots</name>
+			<url>https://repository.apache.org/content/repositories/snapshots/</url>
+			<snapshots><enabled>true</enabled></snapshots>
+		</pluginRepository>
+	</pluginRepositories>
+
 	<build>
 		<pluginManagement>
 			<plugins>
@@ -36,7 +45,7 @@
 				<!-- Can run the exec-jar as: java -Xmx8g -jar target/pcviz.jar -httpPort=8080 -->
 				<groupId>org.apache.tomcat.maven</groupId>
 				<artifactId>tomcat7-maven-plugin</artifactId>
-				<version>2.1</version>
+				<version>2.3-SNAPSHOT</version>
 				<!-- this is for making an executable jar with the tomcat built-in -->
 				<executions>
 					<execution>
@@ -66,14 +75,6 @@
 						<JAVA_OPTS>-Xms256m -Xmx8g -XX:MaxPermSize=256m</JAVA_OPTS>
 					</systemProperties>
 				</configuration>
-				<dependencies>
-					<dependency>
-						<groupId>org.apache.tomcat</groupId>
-						<artifactId>tomcat-catalina</artifactId>
-						<version>7.0.77</version>
-						<scope>runtime</scope>
-					</dependency>
-				</dependencies>
 			</plugin>
 		</plugins>
 	</build>
@@ -183,6 +184,12 @@
 			<groupId>javax.servlet</groupId>
 			<artifactId>jstl</artifactId>
 			<version>1.2</version>
+		</dependency>
+		<dependency>
+			<groupId>org.apache.tomcat</groupId>
+			<artifactId>tomcat-catalina</artifactId>
+			<version>7.0.77</version>
+			<scope>provided</scope>
 		</dependency>
 
 		<!-- test -->
