+++ src/main/java/org/pathwaycommons/pcviz/service/CancerContextService.java
@@ -1,22 +1,3 @@
-/*
- * Copyright 2013 Memorial-Sloan Kettering Cancer Center.
- *
- * This file is part of PCViz.
- *
- * PCViz is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * PCViz is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with PCViz. If not, see <http://www.gnu.org/licenses/>.
- */
-
 package org.pathwaycommons.pcviz.service;
 
 import flexjson.JSONSerializer;
@@ -26,7 +7,6 @@ import org.pathwaycommons.pcviz.model.CancerStudyDetails;
 import org.pathwaycommons.pcviz.model.PropertyKey;
 import org.pathwaycommons.pcviz.cbioportal.*;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
@@ -39,26 +19,8 @@ public class CancerContextService {
 
     private static final Log log = LogFactory.getLog(CancerContextService.class);
     private CBioPortalAccessor cBioPortalAccessor;
-    private String dataCacheFolder;
 
-    /**
-     * Default Constructor.
-     */
     public CancerContextService() {
-    }
-
-
-    @Value("${cbioportal.cache.folder}")
-    public void setDataCacheFolder(String dataCacheFolder) {
-        this.dataCacheFolder = dataCacheFolder;
-    }
-
-    public String getDataCacheFolder() {
-        return dataCacheFolder;
-    }
-
-    public CBioPortalAccessor getcBioPortalAccessor() {
-        return cBioPortalAccessor;
     }
 
     @Autowired
@@ -69,7 +31,7 @@ public class CancerContextService {
     @Cacheable("cancerContextStudiesCache")
     public String listAvailableCancers() throws IOException {
         JSONSerializer jsonSerializer = new JSONSerializer().exclude("*.class");
-        return jsonSerializer.deepSerialize(getcBioPortalAccessor().getCancerStudies());
+        return jsonSerializer.deepSerialize(cBioPortalAccessor.getCancerStudies());
     }
 
     @Cacheable("cancerContextDetailsCache")
@@ -77,18 +39,17 @@ public class CancerContextService {
         CancerStudyDetails cancerStudyDetails = new CancerStudyDetails();
 
         // Get/set cancer study
-        CBioPortalAccessor portal = getcBioPortalAccessor();
-        CancerStudy cancerStudyById = portal.getCancerStudyById(study);
-        portal.setCurrentCancerStudy(cancerStudyById);
+        CancerStudy cancerStudyById = cBioPortalAccessor.getCancerStudyById(study);
+        cBioPortalAccessor.setCurrentCancerStudy(cancerStudyById);
         cancerStudyDetails.setCancerStudy(cancerStudyById);
 
         // Now use the biggest case set as default
-        CaseList caseListById = portal.getCaseListById(study + "_all");
+        CaseList caseListById = cBioPortalAccessor.getCaseListById(study + "_all");
         if(cancerStudyById != null) {
             if(caseListById.getCases()!=null)
                 cancerStudyDetails.setNumberOfCases(caseListById.getCases().length);
             // Now find out if there is profiles
-            for (GeneticProfile geneticProfile : portal.getGeneticProfilesForCurrentStudy()) {
+            for (GeneticProfile geneticProfile : cBioPortalAccessor.getGeneticProfilesForCurrentStudy()) {
                 if (isCNA(geneticProfile)) cancerStudyDetails.setHasCNA(true);
                 if (isExtendedMutation(geneticProfile)) cancerStudyDetails.setHasMutation(true);
                 if (isZscores(geneticProfile)) cancerStudyDetails.setHasExpression(true);
@@ -117,15 +78,13 @@ public class CancerContextService {
     public HashMap<String, HashMap<String, Double>> loadContext(String studyId, String profiles, String genes) throws IOException {
         HashMap<String, HashMap<String, Double>> context = new HashMap<String, HashMap<String, Double>>();
 
-        CBioPortalAccessor.setCacheDir(getDataCacheFolder());
-        CBioPortalAccessor portal = getcBioPortalAccessor();
-        CancerStudy cancerStudyById = portal.getCancerStudyById(studyId);
-        portal.setCurrentCancerStudy(cancerStudyById);
-        CaseList caseListById = portal.getCaseListById(studyId + "_all");
-        portal.setCurrentCaseList(caseListById);
+        CancerStudy cancerStudyById = cBioPortalAccessor.getCancerStudyById(studyId);
+        cBioPortalAccessor.setCurrentCancerStudy(cancerStudyById);
+        CaseList caseListById = cBioPortalAccessor.getCaseListById(studyId + "_all");
+        cBioPortalAccessor.setCurrentCaseList(caseListById);
 
         ArrayList<GeneticProfile> geneticProfiles = new ArrayList<GeneticProfile>();
-        for (GeneticProfile geneticProfile : portal.getGeneticProfilesForCurrentStudy()) {
+        for (GeneticProfile geneticProfile : cBioPortalAccessor.getGeneticProfilesForCurrentStudy()) {
             if(profiles.contains("cna") && isCNA(geneticProfile)) {
                 geneticProfiles.add(geneticProfile);
 
@@ -139,11 +98,11 @@ public class CancerContextService {
                 geneticProfiles.add(geneticProfile);
             }
         }
-        portal.setCurrentGeneticProfiles(geneticProfiles);
+        cBioPortalAccessor.setCurrentGeneticProfiles(geneticProfiles);
 
         for (String gene : genes.split(",")) {
             try {
-                AlterationPack alterations = portal.getAlterations(gene);
+                AlterationPack alterations = cBioPortalAccessor.getAlterations(gene);
                 if (alterations != null) {
                     alterations.complete(Alteration.ANY);
                     double altered = alterations.calcAlteredRatio(Alteration.ANY);
