+++ src/test/resources/ncbi_test.txt
@@ -1,0 +1,9 @@
+Approved Symbol	Entrez Gene ID
+A1BG	1
+A1BG-AS1	503538
+A1CF	29974
+A2M	2
+A2M-AS1	144571
+A2ML1	144568
+A2ML1-AS1
+A2ML1-AS2
