+++ src/test/resources/spring/test.properties
@@ -24,6 +24,7 @@ cocitation.min.node=0
 # HGNC util
 hgnc.location=hgnc_test.txt
 autoCompleteResult.limit=10
+ncbigene.location=ncbi_test.txt
 
 # CBioPortal
 cbioportal.cache.folder=/tmp/pcviz-test/portal
